Updated bitwise ops to use u32 instead of usize

+++ src/matrix/toroidal_bit_matrix.rs
@@ -38,7 +38,7 @@ impl<T: key::Key> ToroidalBinaryMatrix for ToroidalBitMatrix<T> {
 
         let n_bits = rows * cols;
         let n_bits_per_entry = T::n_bits();
-        let n_storage_entries = n_bits.div_ceil(n_bits_per_entry);
+        let n_storage_entries = n_bits.div_ceil(n_bits_per_entry as usize);
         let storage: Vec<T> = vec![T::zero(); n_storage_entries];
 
         let mut result = Self {
@@ -128,16 +128,16 @@ impl<T: key::Key> ToroidalBitMatrix<T> {
             return Err(MatrixConstructError::EmptyTable());
         }
         let bits_per_t = T::n_bits();
-        let n_bits = rows * cols;
-        if n_bits.div_ceil(bits_per_t) != storage.len() {
+        let n_bits = (rows * cols) as u32;
+        if n_bits.div_ceil(bits_per_t) != (storage.len() as u32) {
             return Err(MatrixConstructError::InvalidStorage());
         }
 
-        let n_bits_in_storage = bits_per_t * storage.len();
+        let n_bits_in_storage = bits_per_t * (storage.len() as u32);
         let n_extra_bits = n_bits_in_storage - n_bits;
         if n_extra_bits > 0 {
             let last_byte = storage.last_mut().unwrap();
-            let bit_mask = T::max_value() << n_extra_bits;
+            let bit_mask = T::max_value() << (n_extra_bits as usize);
 
             *last_byte = *last_byte & bit_mask;
         }
@@ -173,7 +173,7 @@ impl<T: key::Key> ToroidalBitMatrix<T> {
     fn get_element_bit_index_from_canon_index(&self, index: (usize, usize)) -> (usize, usize) {
         let (bit_row, bit_col) = index;
         let flat_bit_idx = self.get_cols() * bit_row + bit_col;
-        let bits_per_t = T::n_bits();
+        let bits_per_t = T::n_bits() as usize;
 
         let element_idx = flat_bit_idx / bits_per_t;
         let bit_idx = flat_bit_idx % bits_per_t;
